Fix crosslist privilege-copying issue.
Previously, `section` was not re-defined in the latter loop which copied
privileges, so the bug would cause it to use the last one found in the former
loop which had searched for crosslists, typically resulting in privileges being
copied from the last section in the section file.  Bring these tasks together
into a single loop instead.