Fetch current academic data for each student.
Obtain enrollment status and graduation semester to determine whether each
student is billable and whether they are a pending graduate.

Add helper to determine current business semester, which is the applicable
semester for business purposes, i.e., the current semester if classes are in
session, and the upcoming semester if they are not.