Normalize privileges for many graduate course numbers.
Due to a policy change effective 2020-01-13, basic privileges and several door
privilegs no longer expire for many graduate course numbers at the end of the
semester and should instead be normalized to match their undergraduate
counterparts, when applicable.  There are, of course, a few exceptions due to
various arrangements over the years.