Add Covid door access for S21 class projects.
Access to A31 requested for S21-99361-A per #008691 and for S21-16376 per
follow-up conversation with instructors.  Usage will be limited to transient
access to stored materials.

Use of A5 requested for S21-16480/53399-A per #008803 for projects.  The CS
Gold "HL A5 covid" group is being updated per WO#8701782 to allow for anytime
access (subject to building hours) beginning 2021-03-22.  Reservations will be
required by policy for metering and tracing.