Don't rely on truthiness of Zoho whitelist/blacklist values.
Zoho API v2 returns these values as strings 'true' and 'false', so we can no
longer rely on the truthiness of these values directly.

This was causing all inactive users to be treated as blacklisted and all active
users to be treated as whitelisted, since all non-empty strings are truthy and
those checks came first.