Passport: Provide A5B access to laser-eligible students
Provision door access to A5B via CS Gold to all students in the Grouper group
which determines this door access.  This is based on login eligibility which
is, in turn, based on a combination of community role and safety trainings.
The slight circular dependency on other ACLMAN outputs is not expected to be a
major issue in practice at this time.

The "HL A5B" door access privilege corresponds to CS Gold patron group 2034,
which will actually confer access to two doors:
  - HL A5B itself, during specified laser cutter access hours only
  - HL A5, its containing room and collaborative space, at all hours

The hours for access to HL A5B proper will be adjusted as laser monitor
staffing allows.  All access remains functionally subject to building hours, of
course.

Faculty and staff (in particular, those without a student record) will need to
be managed manually in CS Gold until further code is written to make ACLMAN
better handle non-students by using LDAP.  Additionally, those with Work Alone
permission should be added to a separate CS Gold group which does not have the
same time restrictions as the main group.