Covid: Add S22 door access for A4
Although university-wide access to HL A4 is intended under normal operation, in
late Fall 2021 and over the 2021-2022 winter break, many patrons have routinely
flouted the universal masking policy in this space to an extent that exceeds
monitoring and enforcement capability.

Add special access to allow the S22 classes which meet in this space to
continue accessing it while the broader access policy is temporarily restricted
under WO#9068434.