Add summer-restricted door privs
The normal groups in the CS Gold system for most classrooms have been converted
to suspend normal access over the summer.  Special groups have been created to
override this restriction, with the expectation that the privileges will be set
in the section file for only the relevant summer period.

Use this group to provide access for Summer 2022 classes.